Chatbot component integration with backend
